+++ mysql-learner/tiny-qq/model/friend.go
@@ -1,11 +1,11 @@
 package model
 
 type ParamModifyFriend struct {
-	UserID   int `form:"user_id" json:"user_id"`
-	FriendID int `form:"friend_id" json:"friend_id"`
+	UserID   int64 `form:"user_id" json:"user_id"`
+	FriendID int64 `form:"friend_id" json:"friend_id"`
 }
 
 type UserFriend struct {
-	UserID   int    `json:"user_id"`
+	UserID   int64  `json:"user_id"`
 	Username string `json:"username"`
 }
